Change from long to int as the type to be consistent with Java examples.

+++ pi_cpp_boostMPI.cpp
@@ -1,7 +1,7 @@
 /*
  *  A C++ program to calculate Pi using quadrature as anMPI-based algorithm.
  *
- *  Copyright © 2009-10 Russel Winder
+ *  Copyright © 2009--2011 Russel Winder
  */
 
 #include <iostream>
@@ -10,18 +10,18 @@
 #include "microsecondTime.h"
 
 int main ( int ac , char * * av ) { // MPI requires writeable access to these parameters :-(
-  const long n = 1000000000l ;
+  const int n = 1000000000 ;
   const double delta = 1.0 / n ;
   const long long startTimeMicros = microsecondTime ( ) ;
   boost::mpi::environment environment ( ac , av ) ;
   boost::mpi::communicator world ;
   const int nProcessors = world.size ( ) ;
   const int myId = world.rank ( ) ;
-  const long sliceSize = n / nProcessors ;
-  const long start = 1 + myId * sliceSize ;
-  const long end = ( myId + 1 ) * sliceSize ;
+  const int sliceSize = n / nProcessors ;
+  const int start = 1 + myId * sliceSize ;
+  const int end = ( myId + 1 ) * sliceSize ;
   double localSum = 0.0 ;
-  for ( long i = start ; i <= end ; ++i ) {
+  for ( int i = start ; i <= end ; ++i ) {
     const double x = ( i - 0.5 ) * delta ;
     localSum += 1.0 / ( 1.0 + x * x ) ;
   }
