+++ pi_cpp_sequential.cpp
@@ -1,7 +1,7 @@
 /*
  *  A C++ program to calculate Pi using quadrature as a sequential algorithm.
  *
- *  Copyright © 2008-10 Russel Winder
+ *  Copyright © 2008--2011 Russel Winder
  */
 
 #include <iostream>
@@ -9,11 +9,11 @@
 #include "microsecondTime.h"
 
 int main ( ) {
-  const long n = 1000000000l ;
+  const int n = 1000000000 ;
   const double delta = 1.0 / n ;
   const long long startTimeMicros = microsecondTime ( ) ;
   double sum = 0.0 ;
-  for ( long i = 1 ; i <= n ; ++i ) {
+  for ( int i = 1 ; i <= n ; ++i ) {
     const double x = ( i - 0.5 ) * delta ;
     sum += 1.0 / ( 1.0 + x * x ) ;
   }
