+++ processSlice_cpp.cpp
@@ -1,18 +1,18 @@
 /*
  *  A C++ function to calculate a slice of the overall calculation of Pi using quadrature.
  *
- *  Copyright © 2009-10 Russel Winder
+ *  Copyright © 2009--2011 Russel Winder
  */
 
 extern "C" {
-  double processSlice ( long id , long sliceSize , double delta ) ;
+  double processSlice ( int id , int sliceSize , double delta ) ;
 }
 
-double processSlice ( long id , long sliceSize , double delta ) {
-  const long start = 1 + id * sliceSize ;
-  const long end = ( id + 1 ) * sliceSize ;
+double processSlice ( int id , int sliceSize , double delta ) {
+  const int start = 1 + id * sliceSize ;
+  const int end = ( id + 1 ) * sliceSize ;
   double sum = 0.0 ;
-  for ( long i = start ; i <= end ; ++i ) {
+  for ( int i = start ; i <= end ; ++i ) {
     const double x = ( i - 0.5 ) * delta ;
     sum += 1.0 / ( 1.0 + x * x ) ;
   }
